# apidog.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EvenIt Backend API",
    "version": "1.0.0",
    "description": "Role-based event marketplace backend for auth, onboarding, listings, bookings, subscriptions, uploads, and admin moderation."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ServiceEntity": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "information": {
            "type": "object"
          },
          "pricing": {
            "type": "object"
          },
          "settings": {
            "type": "object"
          },
          "media": {
            "type": "object"
          },
          "availability": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            }
          },
          "publishStatus": {
            "type": "string",
            "enum": [
              "pending",
              "published",
              "rejected"
            ]
          },
          "approvedBy": {
            "nullable": true,
            "type": "object"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviews": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ServiceListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              },
              "hasNextPage": {
                "type": "boolean"
              },
              "hasPrevPage": {
                "type": "boolean"
              }
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceEntity"
            }
          }
        }
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ServiceEntity"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/service-provider/my-services": {
      "get": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Get own services",
        "description": "Authenticated service providers can view only their own services and optionally filter by publish status such as pending or published.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "in": "query",
            "name": "publishStatus",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "published",
                "rejected"
              ]
            },
            "description": "Use `pending` to see requested services awaiting approval, `published` to see approved services, or omit it to see both."
          }
        ],
        "responses": {
          "200": {
            "description": "Own service list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/service-provider/services": {
      "get": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Get published services (public)",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated published services",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Create a service",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Submit multipart form-data. Put the service JSON in the `payload` field and attach optional image files in `images` or `image`. The backend uploads those files and appends the resulting URLs to `media.galleryImages`.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "description": "JSON string matching ServiceCreateRequest. Include `media.galleryImages` only for already-hosted image URLs. Files sent in `images` or `image` are uploaded by the backend automatically.",
                    "example": "{\"information\":{\"serviceName\":\"Premium Catering\",\"category\":\"Catering\",\"description\":\"Corporate and wedding catering service.\",\"serviceArea\":[\"Dhaka\",\"Gazipur\"],\"tags\":[\"wedding\",\"corporate\"]},\"pricing\":{\"amount\":50000,\"pricingType\":\"package\",\"currency\":\"BDT\",\"discount\":{\"type\":\"percentage\",\"value\":10}},\"settings\":{\"amenities\":{\"deliveryIncluded\":true,\"setupIncluded\":true},\"capacity\":300},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://www.youtube.com/watch?v=abc123\"},\"availabilityCalendar\":[{\"date\":\"2026-03-20\",\"hours\":[10,11]},{\"date\":\"2026-03-21\",\"hours\":[14,15]}]}"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Optional service gallery images. Up to 10 files."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional single-image field supported for client compatibility."
                  }
                }
              },
              "examples": {
                "validService": {
                  "summary": "Valid service creation payload",
                  "value": {
                    "payload": "{\"information\":{\"serviceName\":\"Premium Catering\",\"category\":\"Catering\",\"description\":\"Corporate and wedding catering service.\",\"serviceArea\":[\"Dhaka\",\"Gazipur\"],\"tags\":[\"wedding\",\"corporate\"]},\"pricing\":{\"amount\":50000,\"pricingType\":\"package\",\"currency\":\"BDT\",\"discount\":{\"type\":\"percentage\",\"value\":10}},\"settings\":{\"amenities\":{\"deliveryIncluded\":true,\"setupIncluded\":true},\"capacity\":300},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://www.youtube.com/watch?v=abc123\"},\"availabilityCalendar\":[{\"date\":\"2026-03-20\",\"hours\":[10,11]},{\"date\":\"2026-03-21\",\"hours\":[14,15]}]}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/v1/service-provider/services/{serviceId}": {
      "get": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Get one published service by id (public)",
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ServiceEntity"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "patch": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Update service by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Submit multipart form-data. Put the partial service JSON in the `payload` field and attach optional image files in `images` or `image`. The backend uploads those files and appends the resulting URLs to `media.galleryImages`.",
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "payload"
                ],
                "properties": {
                  "payload": {
                    "type": "string",
                    "description": "JSON string matching ServiceUpdateRequest. Files sent in `images` or `image` are uploaded by the backend automatically.",
                    "example": "{\"pricing\":{\"amount\":45000,\"discount\":{\"type\":\"fixed\",\"value\":5000}},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://youtube.com/watch?v=updated-service\"},\"availabilityCalendar\":[{\"date\":\"2026-04-13\",\"hours\":[12]}]}"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Optional service gallery images. Up to 10 files."
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Optional single-image field supported for client compatibility."
                  }
                }
              },
              "examples": {
                "validServiceUpdate": {
                  "summary": "Valid service update payload",
                  "value": {
                    "payload": "{\"pricing\":{\"amount\":45000,\"discount\":{\"type\":\"fixed\",\"value\":5000}},\"media\":{\"galleryImages\":[],\"videoUrl\":\"https://youtube.com/watch?v=updated-service\"},\"availabilityCalendar\":[{\"date\":\"2026-04-13\",\"hours\":[12]}]}"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          }
        }
      },
      "delete": {
        "tags": [
          "ServiceProvider"
        ],
        "summary": "Delete service by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Service deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Service not found"
          }
        }
      }
    },
...
  },
...
}